rorproject master comment and answer links fixed

+++ app/helpers/comments_helper.rb
@@ -1,9 +1,7 @@
 module CommentsHelper
 	def comment_actions(user, comment)
     html = []
-   	
-    
-    if comment.user == user || comment.answer.user == user || user.is_admin?
+    if user.present? && (comment.user == user || comment.answer.user == user || user.is_admin?)
       html << link_to('', comment, method: :delete, :title => "Delete", :class=>'icon-remove', data: { confirm: 'Are you sure?' }, :remote => true)
     end
